refactor: add backend_common to fix lsp
The backend includes only worked by being included in the right
spot inside device_backend.h, which is ugly and breaks tooling
in editors like nvim/vscode LSP. This just moves the required
parts of device_backend into a new backend_common header and
includes it directly from each backend header.
